stbt.{Match,Motion}Timeout: float values are valid for timeout_secs
It was pointed out on the mailing list [1] that the Timeout exception
classes only report integer values for timeout_secs, even if a floating
point value was given. This meants that a timeout from something like:

  wait_for_match("not-going-to-match.png", timeout_secs=0.5)

raised an exception with the message:

  Didn't find match for 'not-going-to-match.png' within 0 seconds.

which is misleading and can lead to confusion as to whether floating point
timeouts are even supported.

[1] https://groups.google.com/forum/#!topic/stb-tester/oHASOHfqh8w