test_save_video: Avoid infinite loop if match fails
This self-test uses `stbt run --save-video video.webm` to create a
video, then verifies the video with a `wait_for_match` against a source
pipeline of "filesrc location=video.webm". For some reason (high CPU
load at the time resulted in a low-quality saved video?) the
`wait_for_match` failed, but the EOS from the end of the video would
cause stbt to restart the source pipeline, thus playing the video again
and restarting the timestamps in such a way that the `wait_for_match`
timeout was never triggered.

Longer term, we want to use some GStreamer mechanism to relate the
buffer's timestamp to the system (wall clock) time. We also want to
avoid restarting the source pipeline for non-live sources. For now this
commit will prevent this self-test from going into an infinite loop if
it fails. There are no other self-tests using filesrc with a video.