Add scripts for creating Ubuntu PPA with our dependencies
When we migrated to GStreamer 1.0 we had to patch some regressions
which had made their way into GStreamer since 0.10.  These patches
have been submitted upstream but not yet made their way into a
stable release of GStreamer, let alone the major distros.  This
makes it very difficult for our users to install and deploy
stb-tester.  This is an attempt at simplifying this process by
providing appropriately patched versions of our dependencies in an
Ubuntu PPA.

This also includes a dkms package to allow distributing a patched
version of the uvcvideo driver.  This is required for Smart TV
support with the Logitech C920 driver.

The code isn't beautiful but it lives in extra/ and is a tools
for maintainers rather than users or developers.