PPA Support: disable `make check` as a part of building Debian package
Disable `make check` for now as it gives us difficulty on Ubuntu's PPA
build farm.  Specifically the lack of a terminal prevents expect from
working and there are timing issues due to the load the build machines are
under.

This commit should be reverted once these issues have been investigated
and resolved but for now PPAs are more important.