Add automated test involving video capture hardware
This is a nascent attempt to write some automated tests to ensure that
stb-tester continues to work with all our current supported hardware.  It's
currently very basic but works.  The intention is that in the future we
will have a test rig such that all commits are tested with real hardware
to prevent future regressions.

The test is currently very basic, it just checks to see if something major
goes wrong but in the future we can add greater control over the input
to the hardware (e.g. using a Raspberry Pi) to measure other properties
such as quality, latency, startup/shutdown time, ability to recover from
video loss, ability to handle HDMI renegotiation, etc.