make publish-copr: New target to publish Fedora RPMs
COPR is a Fedora packaging service, similar to Ubuntu's PPAs.
See http://copr.fedoraproject.org/coprs/stbt/stb-tester/ for details.

COPR is simpler and easier to use than OBS (OpenSUSE Build Service)
which we used previously. All you have to do is upload a Source RPM
somewhere and submit the URL to COPR. COPR is officially supported by
Fedora so it seems a better choice -- if it becomes as widespread as
PPAs for Ubuntu, future stb-tester users will already be familiar with
COPR.

The Dockerfile added in this commit is to help me build the Source RPM
without a host Fedora system. The Vagrantfile is to help me run Docker
without a host Linux system.