Region: Add convenience methods
Sometimes it's convenient to think in terms of width and height, other
times in terms of extents.  The addition of `right`, `bottom` and
`from_extents` to `Region` allow either model to be used as convenient.

These and the `_bounding_box` helper function will be useful for more
advanced uses of OCR which requires manipulating regions.