Refactoring: Factor out loading templates
In preparation for adding `stbt.match`.

Internally we now pass around an `_AnnotatedTemplate` object rather than
an `numpy.ndarray` so we can attach the template name and any other
information we wish to. This tidies up an number of bits of code and
should make them more robust to changes in the future.