stbt-templatematch: use sensible, canonical variable names
This commit attempts to make the variable names used for the two image
file inputs and their corresponding numpy arrays, which previously were
mixed up and confusing.

For example, the file name of the template was stored in the variable
`template_image`, and then the image itself stored in `template`. Yet
`image` is the variable containing the `full_frame` numpy array, and was
given as the value to MatchResult.frame, and the not-yet used but optional
parameter MatchResult.image would (theoretically) be `template_image`.
Most confusing is the call to `stbt._match`, which had `image`, `template`
and `template_image` as arguments!

To make this much less confusing, I've adopted the naming convention used
by stbt templatematch's verbose debug output, where the input frame from
the video source is called `source` and the `template` is called `template`
and further clarified them with `_image` and `_file` suffixes.