stbt batch summary report: Better highlighting of selected row
Instead of colouring the selected row blue, make it the same colour it
already is (green/red/yellow for pass/fail/error) but brighter, and
surrounded by a thick border. This border is open at the right-hand end,
with the row's background colour fading to white, to visually connect
the selected table row with the detailed results pane on the right-hand
side of the page.

We determined the new colours for the selected row by converting the
existing colour to the HSB colour-space and increasing the Brightness to
100%.

Various paddings are reduced (from the values set by Bootstrap) by the
size of the border to keep the same total box size. Note that there was
already a 1px top border on "td"s. "border-collapse: separate" on the
parent table is necessary to avoid a thin border at the right-hand end
of the selected row.