GST_DEBUG is configurable with STBT_CONFIG_FILE
For some infrequently occurring hardware faults, it is useful to have
gstreamer debugging information, for example when Blackmagic cards lock
up. This commit allows users to specify a gstreamer debug string † which
can propagate to stbt run, tv, record, etc. without having to specify it
on the command line or via permanently exported variables (e.g. from
bashrc, etc.)

By setting `gst.gst_debug` in their STBT_CONFIG_FILE, users will get a
"stbt-gst-debug.log" file as an artifact in the working directory, i.e.
the same as "screenshot.png" and "video.webm".

† See https://developer.gnome.org/gstreamer/stable/gst-running.html#GST_DEBUG
An example would be
gst_debug = 5,decklinksrc:6
which sets the debug level to 5 for all elements except decklinksrc (which
is the Blackmagic source element), which has a debug level of 6.