OCR: Return utf-8 object rather than string
Tesseract outputs utf-8 encoded text files.  This allows it to represent
characters outside of 7-bit ASCII such as "£", ligatures, etc.  Up till
now stbt.ocr has also returned strings with the content encoded in utf-8
within.  This is unhelpful if you want to compare against known strings
in a test script.  For example, previously:

    assert '£100' == stbt.ocr()

Would fail as '£100' cannot be represented in 7-bit ASCII.  and:

    assert u'£100' == stbt.ocr()

Would fail as stbt.ocr would return '\xc2\xa3100'.

With this change `stbt.ocr()` would now return u'£100' which is clearer
and simpler to deal with.  This is technically an API break but I don't
expect it to break any real-life test cases.