OCR: Report region searched within in debugging
This is particularly useful if the region is not hard-coded in the test
scripts but generated on the fly from what can be seen on screen.

I've had to disable pylint's E1103 as it was giving false positives with
pylint 0.25 on Travis but not with pylint 1.1.0 on my machine.  The errors
were of the form:

    E1103:715,21:_tesseract: Instance of 'Region' has no 'y' member (but
                             some types could not be inferred)

Which sounds pretty false-positive prone to my ear anyway.