stbt.py: Make all logging output UTF-8
`stbt.ocr()` returns `unicode` strings objects and prints what it reads
to the logs using debug.  This in turn goes to `sys.stderr`.  If OCR has
read any characters outside of the ASCII subset of Unicode (the default
encoding with Python 2) calling `debug` could cause a
`UnicodeEncodeError`[^1].

With this commit all logging output is written as UTF-8.  I think this is
reasonable as it may be consumed by machine rather than a human and we
are unlikely to translate these debugging messages.  The same is not true
of the uses of `sys.stderr` still in stbt.py as they are interactive and
intended to be read by the user.

[^1]: By default whether you get a `UnicodeEncodeError` depends on the
      environment, whether stderr is connected to a TTY amongst other
      things.  So you could get a `UnicodeEncodeError` when writing to a
      file, but not at the terminal!