Refactor: _tesseract: Allow passing a `Gst.Sample` as frame
This was originally written to allow getting the timestamp of frames being
passed to `_tesseract` allowing returning richer information.  This has not
yet been done but being able to pass a `GstSample` is still a nice thing to
be able to do.  It makes _tesseract more like the other image analysis
functions and has the pleasant side-effect of reducing copying.