stbt batch summary report: Default filter from querystring
Anything in the URL's querystring (the part after "?") when the page is
loaded is put into the "filter" search box. Similarly, when you type
into the filter, the querystring is automatically updated to match.

This makes it easy to share links to a subset of results.

This requires using an internal undocumented function of tablequeryjs.
Since we specify the tablequeryjs version explicitly, I can live with
that. We can revisit once a solution for
https://github.com/asimihsan/tablequeryjs/issues/18 is implemented;
in the meantime this functionality is too useful not to merge.