OCR: Add script for measuring OCR effectiveness
validate-ocr.py can be run on a corpus of test images reporting how good a
job stbt has done of reading the text.  Thus it can be used to measure
improvements to the OCR algorithm and, more importantly, detect any
regressions introduced.

This tool is designed such that it can be run on corpuses outside the
stb-tester git tree to allow corpuses containing screen captures from many
set-top boxes without bloating the main stb-tester repo or risking
upsetting the owners of the creators of the various set-top box UIs.

It is not currently run in-tree as a part of make check.

Thanks to david@rothlis.net and lewis.haley@youview.com for review and
fixups.