make check: Enforce pep8 `import` order
PEP8 says: [1]

> Imports should be grouped in the following order:
>
> 1. standard library imports
> 2. related third party imports
> 3. local application/library specific imports
>
> You should put a blank line between each group of imports.

If we're going to follow this convention, let's make the computer check
it to avoid wasting reviewer cycles.

Unfortunately neither the `pep8` tool nor `pylint` have checks to
enforce this. Instead, we use `isort`[2] from PyPI (pip install isort).

If you have trouble getting a file to pass isort's check, run
`isort` (without `--check-only`) on the file to fix it in-place.
To keep this as convenient as possible, don't customise isort's
formatting rules in our pylint.sh.

[1] http://legacy.python.org/dev/peps/pep-0008/#imports
[2] https://pypi.python.org/pypi/isort/