OCR: Add ability to pass configuration to tesseract
tesseract has many configuration values controlling many aspects of the
OCR algorithm.  For instance `tessedit_char_whitelist` can be used to
only recognise certain charactors such as numbers and
`language_model_penalty_non_dict_word` controls the strength of the
dictionary.  Depending on the situation the user can use these to tailor
tesseract's algorithms to improve OCR accuracy.  This commit directly
exposes this configuration functionality/complexity to the user such that
they can set any configuration value.

Ideally we wouldn't expse this level of complexity to the user, but after
experimentation it seems that OCR is an inherently complex field, so this
commit directly exposes the controls that tesseract provide.  This is
better than us adding an additional layer of abstraction/obfuscation in
an attempt to make it simpler for the user.  Once we learn more about the
circumstances under which each configuration value is useful we can
publish recommendations in the documentation.