OCR: Add ability to pass user_patterns and user_words to tesseract
tesseract uses a dictionary to help choose the correct words from the
list of possibilities it produces.  This is helpful when reading real
words but can get in the way of reading serial numbers.

This commit allows the user to help tesseract by giving it hints about
the words it's reading.  This takes the form of adding words to
tesseract's dictionary and adding patterns for tesseract to recognise.

Unfortunately the pattern language is specific to tesseract, and while
it bears some resemblance to regexs the impedance mismatch between the
tesseract pattern language and regex is high enough that a mapping would
be impractical.  As a result the pattern language is exposed verbatim.
Documentation can be found in the tesseract source code[1].

[1]: https://code.google.com/p/tesseract-ocr/source/browse/trunk/dict/trie.h?spec=svn894&r=894#188