extra/vm: Install stb-tester from PPA
Instead of building it from source. This results in a more standardised
configuration, rather than whatever version of the stb-tester source
code the user happened to have checked out. It also gives us one more
way of testing the binary package for Ubuntu.

We still mount the stb-tester source folder in case the user is an
stb-tester developer who wants to run the selftests, install a
development version, etc.  For the same reason we still install
stb-tester's dev dependencies in setup.sh; we don't need to list
the run-time dependencies because they will be installed by
`apt-get install stb-tester`.