stbt-control: Remove unnecessary runtime dependency on nose
The functionality in stbt-control doesn't require nose, but it's tests
did.  Unfortunately as we were using `nose.tools.raises` to decorate
a top level function so we had to import `nose.tools.raises` at the top
level intoducing an unnecessary dependency even when not running the
tests.

This is a little less declarative but is still quite readable and seems to
be better than the alternatives such as moving the test into a different
file or conditionally defining the test function.