Ctrl-C Fixes: Support Ubuntu 12.04 for Travis
PyGObject <3.7.2 doesn't expose the "new" constructor we'd like to be
using, so fall back to __init__.  This means Ctrl-C is broken on 12.04 and
threading will behave differently on Travis than on our supported systems.

This is a real concern because it means Travis isn't testing the actual
implementation that we are using in production.  Hopefully [Travis will
upgrade][1] soon.  Failing that perhaps other CI systems should be
considered.

[1]: https://github.com/travis-ci/travis-ci/issues/2046