Debian: Improve determinism of Debian package generation
Ideally, given a particular commit running `make debian-src-pkg` would
generate a byte-for-byte identical debian package independent of when
and where it was run.  This would have helped to avoid issues uploading
the source packages to the PPA to be built for multiple different
Debian versions where it complains if the source package is not byte-
for-byte identical.

This commit at least takes the time stored in the changelog out of the
equation.