Debian: Build seperate packages for each Debian/Ubuntu distribution
The Debian repository model cannot serve multiple binary packages which
have all been generated from a single source package.  This means that we
need to create a source package for each Ubuntu release so we end up with
binary packages with different names to each other.

This commit adds the Ubuntu release name to the end of the release field
so the binary deb filenames don't clash.  e.g. we end up with:

    stb-tester_0.19-1_amd64.deb
    stb-tester_0.19-1~saucy_amd64.deb
    stb-tester_0.19-1~trusty_amd64.deb

So the target `make ppa-publish` will now publish packages for both
Ubuntu saucy (13.10) and Ubuntu trusty (14.04).