stbt: add sleep() function
This function wraps calls to time.sleep and draws text to the output
video that indicates how long remains until STBT will stop sleeping.

The function is useful because any "plain" call to `time.sleep` will
make the output video seem like STBT is doing nothing (which it is),
but the text on screen makes this a lot more obvious.