stbt control: Allow specifying keymap file in stbt config
With this change the keymap file will default to the value stored in
"control.keymap" rather than and falling back to
`~/.config/stbt/control.conf`.  This makes it easier for the use-case
where you have a config file per device under test and you want to be
able to easily switch between them using the `STBT_CONFIG_FILE`
environment variable.