is_screen_black: Save debug images regardless of outcome
This debugging was added to debug issues when the screen seems black but
`is_screen_black` reports non-black; it seems possible that you might
want to debug an issue where the screen *doesn't* look black but
`is_screen_black` falsely reports it as black.

This is consistent with the debug logging of `detect_match` and
`detect_motion`, which dump debug images regardless of outcome.